glide3x/sst1: replaced several [unsigned] long type usage with FxI32/FxU32
should fix issue noted at
https://github.com/SuperIlu/glide/commit/e8f60ad8bb675a7a61895397706661e7a65bfad2#commitcomment-40186242

the glide2x tree in whole needs a similar massage..

+++ glide3x/sst1/glide3/tests/test11.c
@@ -211,7 +211,7 @@ int main( int argc, char **argv) {
     return 0;
 }
 
-static unsigned long randx = 1;
+static unsigned int randx = 1;
 
 static unsigned int iRandom (unsigned int maxr)
 {
